Moved izgbs and evaluate_location into their own files (#7)
